Add end-to-end tests.
Nose was causing one of them to fail, so instead of trying to figure
that out I decided to just get rid of the Nose dependency. This involved
renaming the test modules so they start with "test_" rather than end with
"_test", since that's the pattern expected by the unittest discover
command.